Mark as end-of-life
This application was not yet marked end-of-life because I had the hope that I would find energy to continue to develop this application. However the reality is that I am not developing this application. Let's explicitly mark it as end of life.

This is what I wrote in 2021 about the current state:

> I stopped development on this project. I was fun to create this application, but I don't have a usecase for my camera anymore. Therefore I have lost interest in adding new features.
>
> I also found out that most bugs reported are from cameras that don't comply to the ONVIF specification. As I don't have such camera available it is not possible to fix that issue. This left multiple issue unsolved. This is not motivating me.
>
> Also I had difficulties releasing this as full free software, because of the non-free license of the ONVIF specification itself.